Miraculously, I got it to work the way I want it to work!
I wasn't instancing a QuincyItem to be able to have its TextureRect exist. I am now doing that, and implemented checking for valid item types. I also updated the Themes for all the TextureRects. They should now be correct.

There is currently one bug that eludes us, and that is how it is possible to click an item that is currently being held if you move to the bottom right quickly enough. It happens accidentally when moving items to the right or down and slightly to the right.